+++ test/helpers.jl
@@ -14,8 +14,8 @@
         @test PT.set_max_print_failures(5) == 0 # returns previous value
         @test PT.MAX_PRINT_FAILURES[] == 5
         @test PT.set_max_print_failures(nothing) == 5 
-        @test PT.MAX_PRINT_FAILURES[] == typemax(Int64)
-        @test PT.set_max_print_failures() == typemax(Int64)
+        @test PT.MAX_PRINT_FAILURES[] == typemax(Int)
+        @test PT.set_max_print_failures() == typemax(Int)
         @test PT.MAX_PRINT_FAILURES[] == 10 # default is 10
     end
 end
